requirements.txt
cffi==1.15.1
cryptography==40.0.2
Django==2.1
mysql-connector-python==8.0.33
Pillow==9.5.0
protobuf==3.20.3
pycparser==2.21
PyMySQL==1.0.3
pytz==2023.3
